WebCamping/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
build==1.2.1
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
colorama==0.4.6
    # via build
django==5.0.6
    # via
    #   -r requirements.in
    #   django-admin-honeypot-updated-2021
    #   django-axes
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-admin-honeypot-updated-2021==1.2.0
    # via -r requirements.in
django-axes==6.4.0
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-cors-middleware==1.5.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-ipware==7.0.1
    # via
    #   -r requirements.in
    #   django-admin-honeypot-updated-2021
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
googlemaps==4.10.0
    # via -r requirements.in
idna==3.7
    # via
    #   -r requirements.in
    #   requests
markdown==3.6
    # via -r requirements.in
packaging==24.0
    # via build
psycopg2==2.9.9
    # via -r requirements.in
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyproject-hooks==1.1.0
    # via build
python-ipware==3.0.0
    # via
    #   -r requirements.in
    #   django-ipware
requests==2.32.1
    # via
    #   -r requirements.in
    #   googlemaps
sqlparse==0.5.0
    # via
    #   -r requirements.in
    #   django
tzdata==2024.1
    # via
    #   -r requirements.in
    #   django
urllib3==2.2.1
    # via
    #   -r requirements.in
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
